requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.6
cryptography==35.0.0
DateTime==4.3
Django==3.2.8
djangorestframework==3.12.4
icalendar==4.0.8
idna==3.2
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==2.2.0
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
uuid==1.30
zope.interface==5.4.0
decorator==5.1.0
django-crontab==0.7.1
typing-extensions==3.10.0.2
validators==0.18.2
numpy==1.21.2
pandas==1.3.4
